Broken down the Appbar component into smaller other components. Reduced the unnecessary re renders of the AppBar considerably.
